# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bookings:
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create booking
      description: Create a new service booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingInput"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List bookings
      description: List all bookings
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get booking
      description: Get a booking by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Booking details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
...
components:
  schemas:
...
    CreateBookingInput:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        suburb:
          type: string
        serviceType:
          type: string
        description:
          type: string
        preferredDate:
          type: string
      required:
        - name
        - email
        - phone
        - address
        - suburb
        - serviceType
        - preferredDate
    Booking:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        suburb:
          type: string
        serviceType:
          type: string
        description:
          type: string
        preferredDate:
          type: string
        depositAmount:
          type: number
        stripeSessionId:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - phone
        - address
        - suburb
        - serviceType
        - preferredDate
        - depositAmount
        - status
        - createdAt
...