# lib/api-zod/src/generated/types/createBookingInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateBookingInput {
  name: string;
  email: string;
  phone: string;
  address: string;
  suburb: string;
  serviceType: string;
  description?: string;
  preferredDate: string;
}